backend/requirements.txt
APScheduler==3.10.4
bidict==0.22.1
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
tzlocal==5.2
Werkzeug==3.0.1
wsproto==1.2.0
